Update broadcast button
